backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
PyJWT==2.10.1
bcrypt==5.0.0